gestinemappfull/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.13
pydantic-settings==2.8.1
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
python-multipart==0.0.20
pypdf==5.4.0
Pillow==12.0.0
pytesseract==0.3.13
